Spec 0013 — Consolidate Base/FieldType*.php → FieldTypes/
FieldType interface and FieldTypeManifest concrete join the FieldTypes/
namespace (with the manifest renamed to FieldTypes/Manifest to match the
sibling Auth\Manifest, Addons\Manifest convention). The interface drops
its Interface suffix and moves to Contracts/FieldTypeManifest. The
service provider aliases the concrete to FieldTypeManifestImpl to avoid
collision with Addons\Manifest.
