+++ packages/core/src/FieldTypes/FieldType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Lunar\Core\Base;
+namespace Lunar\Core\FieldTypes;
 
 interface FieldType
 {
