+++ packages/core/src/FieldTypes/ListField.php
@@ -5,7 +5,6 @@ namespace Lunar\Core\FieldTypes;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Support\Arr;
 use JsonSerializable;
-use Lunar\Core\Base\FieldType;
 use Lunar\Core\Exceptions\FieldTypeException;
 
 class ListField implements Arrayable, FieldType, JsonSerializable
