+++ packages/core/src/FieldTypes/Manifest.php
@@ -1,20 +1,13 @@
 <?php
 
-namespace Lunar\Core\Base;
+namespace Lunar\Core\FieldTypes;
 
 use Illuminate\Support\Collection;
+use Lunar\Core\Contracts\FieldTypeManifest;
 use Lunar\Core\Exceptions\FieldTypes\FieldTypeMissingException;
 use Lunar\Core\Exceptions\FieldTypes\InvalidFieldTypeException;
-use Lunar\Core\FieldTypes\Dropdown;
-use Lunar\Core\FieldTypes\File;
-use Lunar\Core\FieldTypes\ListField;
-use Lunar\Core\FieldTypes\Number;
-use Lunar\Core\FieldTypes\Text;
-use Lunar\Core\FieldTypes\Toggle;
-use Lunar\Core\FieldTypes\TranslatedText;
-use Lunar\Core\FieldTypes\YouTube;
 
-class FieldTypeManifest
+class Manifest implements FieldTypeManifest
 {
     /**
      * The FieldTypes available in Lunar.
@@ -58,10 +51,8 @@ class FieldTypeManifest
 
     /**
      * Return the fieldtypes.
-     *
-     * @return Collection
      */
-    public function getTypes()
+    public function getTypes(): Collection
     {
         return $this->fieldTypes->map(fn ($type) => app()->make($type));
     }
