+++ packages/upgrade/src/Rector/LunarSetList.php
@@ -69,9 +69,9 @@ final class LunarSetList
         'Lunar\\Base\\DiscountTypeInterface' => 'Lunar\\Core\\Base\\DiscountTypeInterface',
         'Lunar\\Base\\Enums\\Concerns\\ProvidesProductAssociationType' => 'Lunar\\Core\\Enums\\Concerns\\ProvidesProductAssociationType',
         'Lunar\\Base\\Enums\\ProductAssociation' => 'Lunar\\Core\\Enums\\ProductAssociation',
-        'Lunar\\Base\\FieldType' => 'Lunar\\Core\\Base\\FieldType',
-        'Lunar\\Base\\FieldTypeManifest' => 'Lunar\\Core\\Base\\FieldTypeManifest',
-        'Lunar\\Base\\FieldTypeManifestInterface' => 'Lunar\\Core\\Base\\FieldTypeManifestInterface',
+        'Lunar\\Base\\FieldType' => 'Lunar\\Core\\FieldTypes\\FieldType',
+        'Lunar\\Base\\FieldTypeManifest' => 'Lunar\\Core\\FieldTypes\\Manifest',
+        'Lunar\\Base\\FieldTypeManifestInterface' => 'Lunar\\Core\\Contracts\\FieldTypeManifest',
         'Lunar\\Base\\HasThumbnailImage' => 'Lunar\\Core\\Base\\HasThumbnailImage',
         'Lunar\\Base\\LunarUser' => 'Lunar\\Core\\Base\\LunarUser',
         'Lunar\\Base\\MediaDefinitionsInterface' => 'Lunar\\Core\\Base\\MediaDefinitionsInterface',
