+++ tests/core/Unit/Base/FieldTypeManifestTest.php
@@ -2,10 +2,10 @@
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Collection;
-use Lunar\Core\Base\FieldTypeManifest;
-use Lunar\Core\Base\FieldTypeManifestInterface;
+use Lunar\Core\Contracts\FieldTypeManifest;
 use Lunar\Core\Exceptions\FieldTypes\FieldTypeMissingException;
 use Lunar\Core\Exceptions\FieldTypes\InvalidFieldTypeException;
+use Lunar\Core\FieldTypes\Manifest;
 use Lunar\Core\Models\Cart;
 use Lunar\Core\Models\Channel;
 use Lunar\Tests\Core\TestCase;
@@ -15,19 +15,19 @@ uses(TestCase::class);
 uses(RefreshDatabase::class);
 
 test('can instantiate class', function () {
-    $manifest = app(FieldTypeManifestInterface::class);
+    $manifest = app(FieldTypeManifest::class);
 
-    expect($manifest)->toBeInstanceOf(FieldTypeManifest::class);
+    expect($manifest)->toBeInstanceOf(Manifest::class);
 });
 
 test('can return types', function () {
-    $manifest = app(FieldTypeManifestInterface::class);
+    $manifest = app(FieldTypeManifest::class);
 
     expect($manifest->getTypes())->toBeInstanceOf(Collection::class);
 });
 
 test('has base types set', function () {
-    $manifest = app(FieldTypeManifestInterface::class);
+    $manifest = app(FieldTypeManifest::class);
 
     expect($manifest->getTypes())->toBeInstanceOf(Collection::class);
 
@@ -35,7 +35,7 @@ test('has base types set', function () {
 });
 
 test('cannot add non fieldtype', function () {
-    $manifest = app(FieldTypeManifestInterface::class);
+    $manifest = app(FieldTypeManifest::class);
 
     $this->expectException(
         InvalidFieldTypeException::class
